fix(estoque): corrige pagina de movimentacoes para build
